+++ app/(site)/not-found.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from "next";
-import "../src/styles/404.scss";
+import "../../src/styles/404.scss";
 
 export const metadata: Metadata = { title: "Page Not Found" };
