+++ app/(site)/pastMeetups/page.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from "next";
-import "../../src/styles/past-meetups-page.scss";
-import HTMLContent from "../_components/Content";
-import MeetupBlock from "../_components/MeetupBlock";
-import { getPastMeetups, getPastMeetupsPage } from "../../lib/content";
+import "../../../src/styles/past-meetups-page.scss";
+import HTMLContent from "../../_components/Content";
+import MeetupBlock from "../../_components/MeetupBlock";
+import { getPastMeetups, getPastMeetupsPage } from "../../../lib/content";
 
 export async function generateMetadata(): Promise<Metadata> {
   const { frontmatter } = await getPastMeetupsPage();
