+++ app/(site)/work-together/page.tsx
@@ -1,0 +1,16 @@
+import type { Metadata } from "next";
+import "../../../src/styles/redirect.scss";
+import RedirectShell from "../../_components/RedirectShell";
+import { getRedirectPage } from "../../../lib/content";
+
+export const metadata: Metadata = { title: "Redirecting..." };
+
+export default async function WorkTogetherPage() {
+  const { frontmatter } = await getRedirectPage("work-together");
+  return (
+    <>
+      <meta httpEquiv="refresh" content={`2;url=${frontmatter.redirectTo}`} />
+      <RedirectShell redirectTo={frontmatter.redirectTo} />
+    </>
+  );
+}
