+++ app/admin/page.tsx
@@ -1,0 +1,15 @@
+import Script from "next/script";
+
+export const metadata = { title: "Content Manager" };
+
+export default function AdminPage() {
+  return (
+    <>
+      <Script
+        type="module"
+        src="https://unpkg.com/@sveltia/cms/dist/sveltia-cms.js"
+        strategy="beforeInteractive"
+      />
+    </>
+  );
+}
